+++ api/python/arrus/system.py
@@ -1,11 +1,13 @@
 import abc
 import dataclasses
 
+
 class SystemCfg(abc.ABC):
     """
     Description of the system with which the user wants to communicate.
 
     """
+    pass
 
 
 @dataclasses.dataclass(frozen=True)
@@ -17,7 +19,7 @@ class CustomUs4RCfg(SystemCfg):
     :param is_hv256: is the hv256 voltage supplier available?
     :param master_us4oem: index of the master Us4OEM module. A master module \
         triggers TxRx execution and should be connected with the voltage \
-        supplier (if available)
+        supplier (if available). By default equal to 0.
     """
     n_us4oems: int
     is_hv256: bool = False